refactor(http): split axum_adapter.rs into handlers module (#316)
Split the session CRUD, stream lifecycle, health/system-stats, and
dictionary route handlers out of the 1891-line axum_adapter.rs into
infrastructure::http::handlers::{sessions,streams,health,dictionary},
grouped by the domain concept each serves. infrastructure::http::dictionary
moves to infrastructure::http::handlers::dictionary as part of the split.

axum_adapter.rs keeps router assembly, PjsAppState, request/response DTOs,
and PjsError/IntoResponse error mapping. PjsAppState's fields become
pub(crate) so the relocated handlers can reach them.

Pure reorganization: route paths, handler behavior, and request/response
shapes are unchanged.
