+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `StreamSession` no longer derives `Debug` (its new `dyn TimeProvider` field can't derive it) — it now has a manual `Debug` impl producing the same field output, with the time-provider field omitted via `finish_non_exhaustive()` (#330)
 - The new `time_provider` field is skipped on `Serialize`/`Deserialize` and restored to `SystemTimeProvider` on deserialize, matching prior behavior (`StreamSession` was never actually serialized with a clock identity to begin with) (#330)
 - Nine timing-sensitive tests in `stream_session_comprehensive.rs` that previously used `std::thread::sleep` to make `Utc::now()` advance (10ms-1100ms delays) now drive a fake `TimeProvider` directly, making those assertions deterministic instead of load-dependent (#330)
+- **BREAKING** `infrastructure::http::axum_adapter.rs` (1891 lines, mixing session CRUD, stream lifecycle, health/metrics, and dictionary route handlers in one file) split its route handler `async fn`s into `infrastructure::http::handlers::{sessions,streams,health,dictionary}`, grouped by the domain concept each serves. `infrastructure::http::dictionary::get_session_dictionary` moved to `infrastructure::http::handlers::dictionary::get_session_dictionary` as part of this split, breaking any external caller importing it by its old path. `axum_adapter.rs` retains router assembly, `PjsAppState`, request/response DTOs, and `PjsError`/`IntoResponse` mapping; `PjsAppState`'s fields are now `pub(crate)` (previously private) so the relocated handlers can reach them. Otherwise a pure reorganization — route paths, handler behavior, and request/response shapes are unchanged (#316)
+- Renamed two unrelated, same-named public enums in `infrastructure::http` that forced callers to disambiguate by fully-qualified path: `axum_extension::StreamError` is now `StreamExtensionError`, and `streaming::StreamError` (including its `infrastructure::http` re-export) is now `StreamTransportError`. No variants or behavior changed (#321)
 
 ### Added
